Move status retrieval to new tools class
Plus small cleanup and fixes.
