+++ .gitignore
@@ -23,3 +23,4 @@ src/test/id*
 # Celery stuff
 celerybeat-schedule.db
 worker_state.db
+celerybeat.pid
