+++ src/cloudlabs/azure_tools.py
@@ -4,6 +4,7 @@ from azure.common.credentials import ServicePrincipalCredentials
 from azure.mgmt.compute import ComputeManagementClient
 from azure.mgmt.resource import ResourceManagementClient
 
+from .host_status import HostStatus
 from .names import group_name, vm_name
 from .secrets import Secrets
 
@@ -52,6 +53,48 @@ class AzureTools(object):
         self.cmc = ComputeManagementClient(self.credentials, self.subscription_id)
         self.rmc = ResourceManagementClient(self.credentials, self.subscription_id)
 
+    def get_status(self, host):
+        """Return the status of a host deployed on Azure."""
+        # Getting the status is not obvious at first glance. This has some info:
+        # https://docs.microsoft.com/en-us/rest/api/compute/virtualmachines/virtualmachines-state
+        # TODO Consider making this static or having a static version
+        try:
+            group = group_name(host)
+            vm = vm_name(host)
+            # The InstanceViewStatus object returned has a code attribute and a
+            # slightly more human-readable display_status. The code seems to follow
+            # a particular structure, so it's probably best to work with that.
+            statuses = [
+                status.code
+                for status
+                in self.cmc.virtual_machines.instance_view(group, vm).statuses
+            ]
+        except Exception as e:
+            return HostStatus.unknown
+        prov_state = _get_provisioning_state_azure(statuses)
+        if prov_state == "deleting":
+            return HostStatus.destroying
+        else:
+            for status in statuses:
+                # Look for the (one?) status which holds the PowerState. This would
+                # be nicer if we could guarantee that there are always exactly two
+                # statuses (provisioning and power), but I'm not sure whether that
+                # is true.
+                if not status.startswith("PowerState"):
+                    continue
+                power_state = _remove_prefix(status, "PowerState/")
+                if power_state == "deallocated":
+                    return HostStatus.stopped
+                elif power_state == "starting":
+                    return HostStatus.starting
+                elif power_state == "running":
+                    return HostStatus.running
+                elif (power_state == "deallocating" or power_state == "stopping"
+                      or power_state == "stopped"):
+                    # "stopped" still incurs charging; "deallocated" means truly off
+                    return HostStatus.stopping
+            return HostStatus.error  # uknown status, return error
+
     def _get_credentials(self):
         """Get the necessary credentials for creating the management clients."""
         subscription_id = Secrets.TF_VAR_azure_subscription_id
@@ -61,3 +104,18 @@ class AzureTools(object):
             tenant=Secrets.TF_VAR_azure_tenant_id
         )
         return credentials, subscription_id
+
+
+def _remove_prefix(string, prefix):
+    # inspired from
+    # https://stackoverflow.com/questions/16891340/remove-a-prefix-from-a-string
+    # but maybe we don't need it
+    return string[len(prefix):] if string.startswith(prefix) else string
+
+
+def _get_provisioning_state_azure(statuses):
+    for status in statuses:
+        if status.startswith("ProvisioningState/"):
+            return _remove_prefix(status, "ProvisioningState/")
+    assert False  # should never happen
+    return None
